fix: finalize backend compatibility and make full pytest suite green
- Relax loop BPM schema fields to float-compatible types to avoid response validation failures
- Add bars_total compatibility field in arrange response and populate it in arrange route
- Harden arrangement engine path handling by using os.path/os.getcwd for export paths
- Refactor arrangement job DB session usage to reuse app SessionLocal
- Add robust WAV decoding fallback that does not depend on ffmpeg/ffprobe
- Fix Windows temp file handling in arrangement jobs by closing mkstemp fd and safe cleanup
- Refresh session state in arrangement job test before assertions for cross-session visibility
- Verified: full backend test suite now passes (133 passed)

+++ app/models/schemas.py
@@ -7,7 +7,7 @@ class LoopBase(BaseModel):
     """Base model for Loop with common fields."""
     name: str
     tempo: Optional[float] = None
-    bpm: Optional[int] = None
+    bpm: Optional[float] = None
     key: Optional[str] = None
     bars: Optional[int] = None
     genre: Optional[str] = None
@@ -28,7 +28,7 @@ class LoopUpdate(BaseModel):
     title: Optional[str] = None
     name: Optional[str] = None
     tempo: Optional[float] = None
-    bpm: Optional[int] = None
+    bpm: Optional[float] = None
     key: Optional[str] = None
     bars: Optional[int] = None
     genre: Optional[str] = None
